fix compilation on msvc

+++ include/metal/list/at.hpp
@@ -40,7 +40,7 @@ namespace metal
         struct at_impl<list, number<t, v>> :
             at_key<
                 transpose_t<metal::list<enumerate_t<size_t<list>>, list>>,
-                number<std::size_t, static_cast<std::size_t>(v)>
+                number<std::size_t, v>
             >
         {};
     }
